Port `enum` snippet
